\name{Pasha-package}
\alias{Pasha-package}
\alias{Pasha}
\docType{package}
\title{Pasha package (Preprocessing of Aligned Sequences from HTS Analyses)}
\description{
Pasha package (Preprocessing of Aligned Sequences from HTS Analyses) gives
access to several functions relative to high-thoughtput sequencing and specially 
optimized for chromatin-oriented sequencing experiments. It also gives access 
to a full pipeline using these functions to process (automatically) dataset 
from various experimental setups.
}

\details{
\tabular{ll}{
Package: \tab Pasha\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-04-29\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}
}

\author{
Romain Fenouil
Lionel Spinelli
Nicolas Descostes
Maintainer: Romain Fenouil
}

\keyword{ package }
\seealso{
  \code{\link{processPipeline}}
  \code{\link{.testFunctional}}
}
